feat(mcp): extract tool list panel logic to separate component #371
Refactor tool list UI into McpToolListPanel for better code organization and reusability. Removes duplicate
