+++ core/src/main/kotlin/cc/unitmesh/devti/mcp/editor/McpPreviewEditor.kt
@@ -3,13 +3,15 @@ package cc.unitmesh.devti.mcp.editor
 import cc.unitmesh.devti.AutoDevIcons
 import cc.unitmesh.devti.llm2.model.LlmConfig
 import cc.unitmesh.devti.llms.custom.CustomLLMProvider
-import cc.unitmesh.devti.mcp.client.CustomMcpServerManager
 import cc.unitmesh.devti.sketch.ui.patch.readText
 import cc.unitmesh.devti.util.AutoDevCoroutineScope
+import com.intellij.openapi.actionSystem.Presentation
+import com.intellij.openapi.actionSystem.impl.ActionButton
 import com.intellij.openapi.application.runReadAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.fileEditor.FileEditor
 import com.intellij.openapi.fileEditor.FileEditorState
+import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.UserDataHolder
 import com.intellij.openapi.util.UserDataHolderBase
@@ -23,14 +25,8 @@ import com.intellij.ui.dsl.builder.Align
 import com.intellij.ui.dsl.builder.panel
 import com.intellij.util.ui.JBUI
 import com.intellij.util.ui.UIUtil
-import com.intellij.openapi.actionSystem.Presentation
-import com.intellij.openapi.actionSystem.impl.ActionButton
-import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.util.ui.components.BorderLayoutPanel
 import io.modelcontextprotocol.kotlin.sdk.Tool
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.cancellable
@@ -38,7 +34,6 @@ import kotlinx.coroutines.launch
 import java.awt.BorderLayout
 import java.awt.Dimension
 import java.awt.FlowLayout
-import java.awt.GridLayout
 import java.beans.PropertyChangeListener
 import javax.swing.*
 import javax.swing.border.CompoundBorder
@@ -51,13 +46,8 @@ open class McpPreviewEditor(
     private var mainEditor = MutableStateFlow<Editor?>(null)
     private val mainPanel = JPanel(BorderLayout())
 
-    private val mcpServerManager = CustomMcpServerManager.instance(project)
-    private val allTools = mutableMapOf<String, List<Tool>>()
-    private var loadingJob: Job? = null
-    private val serverLoadingStatus = mutableMapOf<String, Boolean>()
-    private val serverPanels = mutableMapOf<String, JPanel>()
-
-    private lateinit var toolsContainer: JPanel
+    private var allTools = mutableMapOf<String, List<Tool>>()
+    private lateinit var toolListPanel: McpToolListPanel
     private lateinit var chatbotSelector: JComboBox<String>
     private lateinit var chatInput: JBTextField
     private lateinit var testButton: ActionButton
@@ -65,8 +55,6 @@ open class McpPreviewEditor(
     private lateinit var resultPanel: McpResultPanel
     private val config = McpLlmConfig()
     private val borderColor = JBColor(0xE5E7EB, 0x3C3F41) // Equivalent to Tailwind gray-200
-    private val textGray = JBColor(0x6B7280, 0x9DA0A8)    // Equivalent to Tailwind gray-500
-    private val headerColor = JBColor(0xF3F4F6, 0x2B2D30)  // Light gray for section headers
 
     init {
         createUI()
@@ -75,150 +63,9 @@ open class McpPreviewEditor(
 
     private fun loadTools() {
         val content = runReadAction { virtualFile.readText() }
-        loadingJob?.cancel()
-        serverLoadingStatus.clear()
-        serverPanels.clear()
-        allTools.clear()
-        
-        SwingUtilities.invokeLater {
-            toolsContainer.removeAll()
-            toolsContainer.revalidate()
-            toolsContainer.repaint()
+        toolListPanel.loadTools(content) { tools ->
+            this.allTools = tools
         }
-        
-        loadingJob = CoroutineScope(Dispatchers.IO).launch {
-            val serverConfigs = mcpServerManager.getServerConfigs(content)
-            
-            if (serverConfigs.isNullOrEmpty()) {
-                SwingUtilities.invokeLater {
-                    showNoServersMessage()
-                }
-                return@launch
-            }
-            
-            SwingUtilities.invokeLater {
-                serverConfigs.keys.forEach { serverName ->
-                    serverLoadingStatus[serverName] = true
-                    createServerSection(serverName)
-                }
-            }
-            
-            serverConfigs.forEach { (serverName, serverConfig) ->
-                try {
-                    val tools = mcpServerManager.collectServerInfo(serverName, serverConfig)
-                    allTools[serverName] = tools
-                    
-                    SwingUtilities.invokeLater {
-                        updateServerSection(serverName, tools)
-                        serverLoadingStatus[serverName] = false
-                    }
-                } catch (e: Exception) {
-                    SwingUtilities.invokeLater {
-                        showServerError(serverName, e.message ?: "Unknown error")
-                        serverLoadingStatus[serverName] = false
-                    }
-                }
-            }
-        }
-    }
-    
-    private fun createServerSection(serverName: String) {
-        val serverPanel = JPanel(BorderLayout()).apply {
-            background = UIUtil.getPanelBackground()
-            border = BorderFactory.createCompoundBorder(
-                BorderFactory.createMatteBorder(0, 0, 1, 0, borderColor),
-                JBUI.Borders.empty(4, 0)
-            )
-        }
-        
-        val headerPanel = JPanel(BorderLayout()).apply {
-            background = headerColor
-            border = JBUI.Borders.empty(4, 8)
-        }
-        
-        val serverLabel = JBLabel(serverName).apply {
-            font = JBUI.Fonts.label(14.0f).asBold()
-            foreground = UIUtil.getLabelForeground()
-        }
-        
-        headerPanel.add(serverLabel, BorderLayout.WEST)
-        serverPanel.add(headerPanel, BorderLayout.NORTH)
-        
-        val toolsPanel = JPanel(GridLayout(0, 3, 4, 4)).apply {
-            background = UIUtil.getPanelBackground()
-            border = JBUI.Borders.empty()
-        }
-        
-        val loadingLabel = JBLabel("Loading tools from $serverName...").apply {
-            font = JBUI.Fonts.label(12.0f)
-            foreground = textGray
-            horizontalAlignment = SwingConstants.LEFT
-            icon = AutoDevIcons.LOADING
-            iconTextGap = JBUI.scale(8)
-        }
-        
-        toolsPanel.add(loadingLabel)
-        serverPanel.add(toolsPanel, BorderLayout.CENTER)
-        
-        serverPanels[serverName] = toolsPanel
-        
-        toolsContainer.add(serverPanel)
-        toolsContainer.revalidate()
-        toolsContainer.repaint()
-    }
-    
-    private fun updateServerSection(serverName: String, tools: List<Tool>) {
-        val toolsPanel = serverPanels[serverName] ?: return
-        toolsPanel.removeAll()
-        
-        if (tools.isEmpty()) {
-            val noToolsLabel = JBLabel("No tools available for $serverName").apply {
-                foreground = textGray
-                horizontalAlignment = SwingConstants.LEFT
-            }
-            toolsPanel.add(noToolsLabel)
-        } else {
-            tools.forEach { tool ->
-                val panel = McpToolDetailPanel(project, serverName, tool)
-                toolsPanel.add(panel)
-            }
-        }
-        
-        toolsPanel.revalidate()
-        toolsPanel.repaint()
-    }
-    
-    private fun showServerError(serverName: String, errorMessage: String) {
-        val toolsPanel = serverPanels[serverName] ?: return
-        toolsPanel.removeAll()
-        
-        val errorLabel = JBLabel("Error loading tools: $errorMessage").apply {
-            foreground = JBColor.RED
-            horizontalAlignment = SwingConstants.LEFT
-        }
-        
-        toolsPanel.add(errorLabel)
-        toolsPanel.revalidate()
-        toolsPanel.repaint()
-    }
-    
-    private fun showNoServersMessage() {
-        toolsContainer.removeAll()
-        
-        val noServersPanel = JPanel(BorderLayout()).apply {
-            background = UIUtil.getPanelBackground()
-            border = JBUI.Borders.empty(16)
-        }
-        
-        val noServersLabel = JBLabel("No MCP servers configured. Please check your configuration.").apply {
-            foreground = textGray
-            horizontalAlignment = SwingConstants.CENTER
-        }
-        
-        noServersPanel.add(noServersLabel, BorderLayout.CENTER)
-        toolsContainer.add(noServersPanel)
-        toolsContainer.revalidate()
-        toolsContainer.repaint()
     }
 
     fun refreshMcpTool() {
@@ -244,12 +91,9 @@ open class McpPreviewEditor(
             border = JBUI.Borders.empty(4)
         }
 
-        toolsContainer = JPanel().apply {
-            layout = BoxLayout(this, BoxLayout.Y_AXIS)
-            background = UIUtil.getPanelBackground()
-        }
-
-        val toolsScrollPane = JBScrollPane(toolsContainer).apply {
+        toolListPanel = McpToolListPanel(project)
+        
+        val toolsScrollPane = JBScrollPane(toolListPanel).apply {
             border = BorderFactory.createEmptyBorder()
             background = UIUtil.getPanelBackground()
         }
@@ -401,7 +245,6 @@ open class McpPreviewEditor(
     override fun addPropertyChangeListener(listener: PropertyChangeListener) {}
     override fun removePropertyChangeListener(listener: PropertyChangeListener) {}
     override fun dispose() {
-        loadingJob?.cancel()
+        toolListPanel.dispose()
     }
 }
-
